SOFTWARE/APP/v4/iva_key_app/requirements.txt
PyQt6
speech_recognition
pynput
pyserial
json
pycaw
keyboard